requirements.txt
asgiref==3.9.2
awsebcli==3.25
bcrypt==5.0.0
blessed==1.22.0
botocore==1.40.41
cement==2.10.14
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.1
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.6
fabric==3.2.2
gunicorn==20.1.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pycparser==2.23
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.14
whitenoise==6.11.0
wrapt==1.17.3
python-dotenv
